requirements.txt
tkinter
customtkinter
pdf2doc